gcf_1/requirements.txt
google-cloud-storage==1.28.1
Wand==0.6.1
